refactor: add feature folder for comments
