+++ src/features/listings/components/ListingPage/index.ts
@@ -1,4 +1,3 @@
-export * from './Comments';
 export * from './DescriptionUI';
 export * from './ListingHeader';
 export * from './ListingWinners';
